fix floating truncation

+++ mvla.h
@@ -2377,18 +2377,18 @@ MVLAIMPL v4f_t v4f_div(v4f_t a, v4f_t b) {
 }
 
 MVLAIMPL v4f_t v4f_min(v4f_t a, v4f_t b) {
-  a.x = fmin(a.x, b.x);
-  a.y = fmin(a.y, b.y);
-  a.z = fmin(a.z, b.z);
-  a.w = fmin(a.w, b.w);
+  a.x = fminf(a.x, b.x);
+  a.y = fminf(a.y, b.y);
+  a.z = fminf(a.z, b.z);
+  a.w = fminf(a.w, b.w);
   return a;
 }
 
 MVLAIMPL v4f_t v4f_max(v4f_t a, v4f_t b) {
-  a.x = fmax(a.x, b.x);
-  a.y = fmax(a.y, b.y);
-  a.z = fmax(a.z, b.z);
-  a.w = fmax(a.w, b.w);
+  a.x = fmaxf(a.x, b.x);
+  a.y = fmaxf(a.y, b.y);
+  a.z = fmaxf(a.z, b.z);
+  a.w = fmaxf(a.w, b.w);
   return a;
 }
